Clean up response traits
