+++ src/body/sync.rs
@@ -20,25 +20,39 @@ use std::{
 pub struct Body(Inner);
 
 enum Inner {
+    Empty,
     Buffer(Cursor<Cow<'static, [u8]>>),
     Reader(Box<dyn Read + Send + Sync>, Option<u64>),
 }
 
 impl Body {
-    pub fn from_reader<R>(reader: R) -> Self
-    where
-        R: Read + Send + Sync + 'static,
-    {
-        Self(Inner::Reader(Box::new(reader), None))
+    /// Create a new empty body.
+    ///
+    /// An empty body represents the *absence* of a body, which is semantically
+    /// different than the presence of a body of zero length.
+    pub const fn empty() -> Self {
+        Self(Inner::Empty)
     }
 
-    pub fn from_reader_sized<R>(reader: R, length: u64) -> Self
-    where
-        R: Read + Send + Sync + 'static,
-    {
-        Self(Inner::Reader(Box::new(reader), Some(length)))
-    }
-
+    /// Create a new body from a potentially static byte buffer.
+    ///
+    /// The body will have a known length equal to the number of bytes given.
+    ///
+    /// This will try to prevent a copy if the type passed in can be re-used,
+    /// otherwise the buffer will be copied first. This method guarantees to not
+    /// require a copy for the following types:
+    ///
+    /// - `&'static [u8]`
+    /// - `&'static str`
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use isahc::Body;
+    ///
+    /// // Create a body from a static string.
+    /// let body = Body::from_bytes_static("hello world");
+    /// ```
     #[inline]
     pub fn from_bytes_static<B>(bytes: B) -> Self
     where
@@ -52,15 +66,74 @@ impl Body {
         }
     }
 
+    /// Create a streaming body that reads from the given reader.
+    ///
+    /// The body will have an unknown length. When used as a request body,
+    /// [chunked transfer
+    /// encoding](https://tools.ietf.org/html/rfc7230#section-4.1) might be used
+    /// to send the request.
+    pub fn from_reader<R>(reader: R) -> Self
+    where
+        R: Read + Send + Sync + 'static,
+    {
+        Self(Inner::Reader(Box::new(reader), None))
+    }
+
+    /// Create a streaming body with a known length.
+    ///
+    /// If the size of the body is known in advance, such as with a file, then
+    /// this function can be used to create a body that can determine its
+    /// `Content-Length` while still reading the bytes asynchronously.
+    ///
+    /// Giving a value for `length` that doesn't actually match how much data
+    /// the reader will produce may result in errors when sending the body in a
+    /// request.
+    pub fn from_reader_sized<R>(reader: R, length: u64) -> Self
+    where
+        R: Read + Send + Sync + 'static,
+    {
+        Self(Inner::Reader(Box::new(reader), Some(length)))
+    }
+
+    /// Report if this body is empty.
+    ///
+    /// This is not necessarily the same as checking for `self.len() ==
+    /// Some(0)`. Since HTTP message bodies are optional, there is a semantic
+    /// difference between the absence of a body and the presence of a
+    /// zero-length body. This method will only return `true` for the former.
+    pub fn is_empty(&self) -> bool {
+        match self.0 {
+            Inner::Empty => true,
+            _ => false,
+        }
+    }
+
+    /// Get the size of the body, if known.
+    ///
+    /// The value reported by this method is used to set the `Content-Length`
+    /// for outgoing requests.
+    ///
+    /// When coming from a response, this method will report the value of the
+    /// `Content-Length` response header if present. If this method returns
+    /// `None` then there's a good chance that the server used something like
+    /// chunked transfer encoding to send the response body.
+    ///
+    /// Since the length may be determined totally separately from the actual
+    /// bytes, even if a value is returned it should not be relied on as always
+    /// being accurate, and should be treated as a "hint".
     pub fn len(&self) -> Option<u64> {
         match &self.0 {
+            Inner::Empty => Some(0),
             Inner::Buffer(bytes) => Some(bytes.get_ref().len() as u64),
             Inner::Reader(_, len) => *len,
         }
     }
 
+    /// If this body is repeatable, reset the body stream back to the start of
+    /// the content. Returns `false` if the body cannot be reset.
     pub fn reset(&mut self) -> bool {
         match &mut self.0 {
+            Inner::Empty => true,
             Inner::Buffer(cursor) => {
                 cursor.set_position(0);
                 true
@@ -82,6 +155,7 @@ impl Body {
     /// blocking fashion.
     pub(crate) fn into_async(self) -> (AsyncBody, Option<Writer>) {
         match self.0 {
+            Inner::Empty => (AsyncBody::empty(), None),
             Inner::Buffer(cursor) => (AsyncBody::from_bytes_static(cursor.into_inner()), None),
             Inner::Reader(reader, len) => {
                 let (pipe_reader, writer) = pipe();
@@ -102,15 +176,22 @@ impl Body {
 impl Read for Body {
     fn read(&mut self, buf: &mut [u8]) -> Result<usize> {
         match &mut self.0 {
+            Inner::Empty => Ok(0),
             Inner::Buffer(cursor) => cursor.read(buf),
             Inner::Reader(reader, _) => reader.read(buf),
         }
     }
 }
 
+impl Default for Body {
+    fn default() -> Self {
+        Self::empty()
+    }
+}
+
 impl From<()> for Body {
     fn from(_: ()) -> Self {
-        Self::from("")
+        Self::empty()
     }
 }
 
@@ -193,5 +274,28 @@ impl Writer {
 
             self.writer.write_all(&buf[..len]).await?;
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    static_assertions::assert_impl_all!(Body: Send, Sync);
+
+    #[test]
+    fn empty_body() {
+        let body = Body::empty();
+
+        assert!(body.is_empty());
+        assert_eq!(body.len(), Some(0));
+    }
+
+    #[test]
+    fn zero_length_body() {
+        let body = Body::from(vec![]);
+
+        assert!(!body.is_empty());
+        assert_eq!(body.len(), Some(0));
     }
 }
